Move grpc protos to a common subdirectory
.. and split out the request and response messages to allow for sharing.
