+++ pkg/protos/services/trace_sink.proto
@@ -1,0 +1,31 @@
+syntax = "proto3";
+
+package services;
+
+import "github.com/Jim3Things/CloudChamber/pkg/protos/log/entry.proto";
+import "github.com/Jim3Things/CloudChamber/pkg/protos/services/requests.proto";
+
+import "google/protobuf/empty.proto";
+
+option go_package = "github.com/Jim3Things/CloudChamber/pkg/protos/services";
+
+// This service is a common collection point for trace entries across
+// Cloud Chamber.
+//
+// This provides a single service that the UI can query to show a single
+// logical flow of action across the Cloud Chamber services.
+service TraceSink {
+    // Append the specified entry to the set of traces held by the trace sink
+    rpc Append(AppendRequest) returns (google.protobuf.Empty);
+
+    // Get the trace entries after the specified trace ID, up to the specified
+    // limit.
+    rpc GetAfter(GetAfterRequest) returns (GetAfterResponse);
+
+    // Get the active policy options in use by the trace sink
+    rpc GetPolicy(GetPolicyRequest) returns (GetPolicyResponse);
+
+    // Reset the trace sink to its initial state.  Used for unit testing.
+    rpc Reset(ResetRequest) returns (google.protobuf.Empty);
+}
+
